SBERDOMA-62 refactor: extract checkOrganizationPermission
