+++ apps/condo/domains/organization/utils/accessSchema.js
@@ -1,0 +1,20 @@
+const { getByCondition } = require('@core/keystone/schema')
+
+async function checkOrganizationPermission (userId, organizationId, permission) {
+    if (!userId || !organizationId) return false
+    const employee = await getByCondition('OrganizationEmployee', {
+        organization: { id: organizationId },
+        user: { id: userId },
+    })
+    if (!employee || !employee.role) return false
+    const employeeRole = await getByCondition('OrganizationEmployeeRole', {
+        id: employee.role,
+        organization: { id: organizationId },
+    })
+    if (!employeeRole) return false
+    return employeeRole[permission] || false
+}
+
+module.exports = {
+    checkOrganizationPermission,
+}
